+++ src/components/sub-components/Device.tsx
@@ -23,14 +23,15 @@ const Device: React.FC<SubComponentProps> = ({ device, updateQuery, setDevice, d
     let allResults = [] as any;
 
     async function getDeviceOptions() {
+      console.log('getDeviceOptions deviceGroup: ', deviceGroup);
       try {
         let token = '';
         token = await datasource.getToken();
-        if (deviceGroup === undefined || deviceGroup === null) {
+        if (deviceGroup === null) {
           results = await datasource.sevOneConnection.getDevices(token, 3, 20, 0, '');
           allResults = await datasource.sevOneConnection.getAllDevices(token, 3);
         } else {
-          allResults = await datasource.sevOneConnection.getDeviceGroupMembers(token, 3, deviceGroup.value, '');
+          allResults = await datasource.sevOneConnection.getDeviceGroupMembers(token, 3, deviceGroup, '');
           results = allResults.slice(0, 20);
         }
       } catch (err) {